requirements.txt
Django==4.0
# DRF
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
markdown==3.3.6
django-filter==21.1
factory-boy==3.2.0
pytest-django==4.5.2
drf-spectacular==0.21.1
django-cors-headers==3.14.0

psycopg2==2.9.1
# django-dotenv==1.4.2
